# Improved InitSampling function speed by 2.12 times