--- a/src/tree/updater_quantile_hist.cc
+++ b/src/tree/updater_quantile_hist.cc
@@ -691,17 +691,17 @@ void QuantileHistMaker::Builder<GradientSumT>::InitSampling(const std::vector<Gr
     r = rnd;
   }
   const size_t discard_size = info.num_row_ / nthread;
+  uint32_t coin_flip_border = static_cast<uint32_t>(std::numeric_limits<uint32_t>::max() * param_.subsample);
   #pragma omp parallel num_threads(nthread)
   {
     const size_t tid = omp_get_thread_num();
     const size_t ibegin = tid * discard_size;
     const size_t iend = (tid == (nthread - 1)) ?
                         info.num_row_ : ibegin + discard_size;
-    std::bernoulli_distribution coin_flip(param_.subsample);
 
-    rnds[tid].discard(2*discard_size * tid);
+    rnds[tid].discard(discard_size * tid);
     for (size_t i = ibegin; i < iend; ++i) {
-      if (gpair[i].GetHess() >= 0.0f && coin_flip(rnds[tid])) {
+      if (gpair[i].GetHess() >= 0.0f && rnds[tid]() < coin_flip_border) {
         p_row_indices[ibegin + row_offsets[tid]++] = i;
       }
     }