Discarding elements from generators takes up most of the working time in `InitSampling`.
Since stdlibc++ doesn't have any random engines with o(n) complexity (little-o), we only can optimize the number of discarded elements.
 `std::bernoulli_distribution` requires 64-bit input, so in previous version we had to discard twice as much elements as now.
This little optimization gives us ~2.12 speed up of `InitSampling` time, which translates into up to 16% speed up of the whole training time when `subsampling < 1`
The quality remains the same:


Mortgage dataset |  |  |
-- | -- | --
version | training time | RMSE
original | 35.479 | 0.009271
optimized | 30.538 | 0.009262


Santander dataset |  |  |
-- | -- | --
version | training time | Log Loss
original | 242.249 | 0.166067
optimized | 229.607 | 0.166101


Plastic dataset |  |  |
-- | -- | --
version | training time | Log Loss
original | 1.206 | 0.093388
optimized | 1.180 | 0.094090